backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# LlamaIndex for RAG (OpenAI only)
llama-index>=0.12.0
llama-index-llms-openai>=0.3.0
llama-index-vector-stores-pinecone>=0.1.0

# Twilio for OTP/SMS
twilio>=8.0.0

# Pinecone Vector Database
pinecone-client>=3.0.0

# PDF processing
pypdf>=4.0.0
pdfplumber>=0.10.0

# LLM provider
openai>=1.30.0

# Environment variables
python-dotenv>=1.0.0

# Database (PostgreSQL)
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0        # Async PostgreSQL driver
psycopg2-binary>=2.9.0 # Sync driver for Alembic
alembic>=1.13.0        # Database migrations

# Authentication
passlib[bcrypt]>=1.7.4 # Password hashing
python-jose[cryptography]>=3.3.0  # JWT tokens
bcrypt>=4.0.0

# Validation
pydantic[email]>=2.0.0
email-validator>=2.0.0

# Caching (optional - for production)
# redis>=5.0.0
# aioredis>=2.0.0
